refactor: move utils tests to a separate module
