+++ tests/utils/test_ephemeral_docker_image.py
